Get rid of dictionary file and just load it directly in the migrate script'
